# web/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lettura API",
    "description": "Lettura is a wallabag-inspired read-it-later service. Endpoints below are stable; unannotated routes exist but their schema is not yet machine-readable.",
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/admin/reindex": {
      "post": {
        "tags": [
          "admin"
        ],
        "operationId": "reindex",
        "responses": {
          "200": {
            "description": "Reindex complete",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReindexResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid auth"
          },
          "403": {
            "description": "Admin role required"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/admin/users": {
      "get": {
        "tags": [
          "admin"
        ],
        "operationId": "list_users",
        "responses": {
          "200": {
            "description": "List of user summaries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserSummary"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid auth"
          },
          "403": {
            "description": "Admin role required"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/annotations/{id}": {
      "delete": {
        "tags": [
          "annotations"
        ],
        "operationId": "delete_annotation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Annotation ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Annotation deleted"
          },
          "401": {
            "description": "Missing or invalid auth"
          },
          "404": {
            "description": "Annotation not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "tags": [
          "annotations"
        ],
        "operationId": "update_annotation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Annotation ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAnnotation"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Annotation updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Annotation"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid auth"
          },
          "404": {
            "description": "Annotation not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/audit-logs": {
      "get": {
        "tags": [
          "audit-logs"
        ],
        "operationId": "list_audit_logs",
        "parameters": [
          {
            "name": "action",
            "in": "query",
            "required": false,
            "schema": {
              "oneOf": [
                {
                  "type": "null"
                },
                {
                  "$ref": "#/components/schemas/AuditAction"
                }
              ]
            }
          },
          {
            "name": "resource_type",
            "in": "query",
            "required": false,
            "schema": {
              "oneOf": [
                {
                  "type": "null"
                },
                {
                  "$ref": "#/components/schemas/AuditResourceType"
                }
              ]
            }
          },
          {
            "name": "resource_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "format": "uuid"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of audit log entries",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListAuditLogsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid auth"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/entries/{entry_id}/annotations": {
      "get": {
        "tags": [
          "annotations"
        ],
        "operationId": "list_annotations",
        "parameters": [
          {
            "name": "entry_id",
            "in": "path",
            "description": "Entry ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Annotations for the entry",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Annotation"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid auth"
          },
          "404": {
            "description": "Entry not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "annotations"
        ],
        "operationId": "create_annotation",
        "parameters": [
          {
            "name": "entry_id",
            "in": "path",
            "description": "Entry ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAnnotation"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Annotation created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Annotation"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid auth"
          },
          "404": {
            "description": "Entry not found"
          },
          "422": {
            "description": "Validation error"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Annotation": {
        "type": "object",
        "required": [
          "id",
          "entry_id",
          "user_id",
          "quote",
          "text",
          "ranges",
          "is_orphaned",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "entry_id": {
            "type": "string",
            "format": "uuid"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "is_orphaned": {
            "type": "boolean"
          },
          "quote": {
            "type": "string"
          },
          "ranges": {},
          "text": {
            "type": "string"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "AuditAction": {
        "type": "string",
        "enum": [
          "register",
          "login",
          "logout",
          "refresh_token",
          "change_password",
          "regenerate_feed_token",
          "create_pat",
          "delete_pat",
          "create_entry",
          "update_entry",
          "soft_delete_entry",
          "restore_entry",
          "permanent_delete_entry",
          "archive_entry",
          "unarchive_entry",
          "star_entry",
          "unstar_entry",
          "refetch_entry",
          "create_tag",
          "delete_tag",
          "add_tag_to_entry",
          "remove_tag_from_entry",
          "create_annotation",
          "update_annotation",
          "delete_annotation",
          "create_memo",
          "delete_memo",
          "promote_memo",
          "create_tagging_rule",
          "update_tagging_rule",
          "delete_tagging_rule",
          "create_site_rule",
          "update_site_rule",
          "delete_site_rule",
          "import_wallabag",
          "import_browser",
          "import_lettura",
          "export_all",
          "create_page",
          "update_page",
          "delete_page",
          "restore_page",
          "admin_backup",
          "admin_restore",
          "admin_reindex",
          "admin_list_users",
          "bulk_tag_add",
          "bulk_untag",
          "bulk_archive",
          "bulk_star",
          "bulk_soft_delete",
          "rename_tag",
          "upload_page_files"
        ]
      },
...
      "AuditLog": {
        "type": "object",
        "required": [
          "id",
          "auth_source",
          "action",
          "status",
          "details",
          "created_at"
        ],
        "properties": {
          "action": {
            "$ref": "#/components/schemas/AuditAction"
          },
          "auth_source": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "details": {},
          "error_message": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "ip_address": {
            "type": [
              "string",
              "null"
            ]
          },
          "request_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "resource_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "resource_type": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AuditResourceType"
              }
            ]
          },
          "status": {
            "type": "string"
          },
          "user_agent": {
            "type": [
              "string",
              "null"
            ]
          },
          "user_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          }
        }
      },
      "AuditResourceType": {
        "type": "string",
        "enum": [
          "user",
          "entry",
          "tag",
          "annotation",
          "memo",
          "tagging_rule",
          "site_rule",
          "page",
          "pat",
          "system"
        ]
      },
...
      "CreateAnnotation": {
        "type": "object",
        "required": [
          "quote",
          "ranges"
        ],
        "properties": {
          "quote": {
            "type": "string"
          },
          "ranges": {},
          "text": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "ListAuditLogsResponse": {
        "type": "object",
        "required": [
          "data",
          "total",
          "limit",
          "offset"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuditLog"
            }
          },
          "limit": {
            "type": "integer",
            "format": "int64"
          },
          "offset": {
            "type": "integer",
            "format": "int64"
          },
          "total": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "ReindexResponse": {
        "type": "object",
        "required": [
          "message",
          "indexed"
        ],
        "properties": {
          "indexed": {
            "type": "integer",
            "minimum": 0
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "UpdateAnnotation": {
        "type": "object",
        "properties": {
          "ranges": {},
          "text": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "UserSummary": {
        "type": "object",
        "required": [
          "id",
          "username",
          "email",
          "is_admin",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "email": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "is_admin": {
            "type": "boolean"
          },
          "username": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}